fix: revert #7100 (master returns rbacEnabled = false...) #7216
This reverts commit 3e76f63e934ec62846ab14fa5270c6d46d32b6a8.
