+++ e2e_tests/pytest.ini
@@ -10,7 +10,6 @@ markers =
     e2e_cpu_2a: end to end CPU tests with two agents
     e2e_cpu_elastic: end to end CPU tests with elasticsearch for logging
     e2e_cpu_postgres: end to end CPU tests for testing basic database functionality
-    e2e_cpu_rbac: end to end ee rbac tests
     e2e_cpu_cross_version: end to end CPU tests for testing basic cluster functionality with differing master/agent versions
     e2e_cpu_agent_connection_loss: end to end CPU tests for testing agent functionality on connection loss
     e2e_gpu: end to end GPU tests
