+++ e2e_tests/tests/api_utils.py
@@ -5,8 +5,7 @@ from determined.common import api
 from determined.common.api import Session, authentication, bindings, certs
 from determined.common.api._util import AnyNTSC, NTSC_Kind
 from tests import config as conf
-
-ADMIN_CREDENTIALS = authentication.Credentials("admin", "")
+from tests.cluster import test_users
 
 
 def get_random_string() -> str:
@@ -21,7 +20,7 @@ def determined_test_session(
 
     if credentials is None:
         if admin:
-            credentials = ADMIN_CREDENTIALS
+            credentials = test_users.ADMIN_CREDENTIALS
         else:
             credentials = authentication.Credentials("determined", "")
