+++ e2e_tests/tests/cluster/managed_cluster_k8s.py
@@ -5,10 +5,10 @@ import pytest
 from kubernetes import client, config, watch
 
 from tests import config as conf
-from tests.utils import det_cmd
 
 from .abstract_cluster import Cluster
 from .managed_cluster import get_agent_data
+from .test_groups import det_cmd
 from .utils import run_command, wait_for_command_state
