+++ e2e_tests/tests/cluster/test_master_restart.py
@@ -13,11 +13,11 @@ from tests import command as cmd
 from tests import config as conf
 from tests import experiment as exp
 from tests.cluster.test_users import det_spawn
-from tests.utils import det_cmd_json
 
 from .abstract_cluster import Cluster
 from .managed_cluster import ManagedCluster, get_agent_data
 from .managed_cluster_k8s import ManagedK8sCluster
+from .test_groups import det_cmd_json
 from .utils import (
     command_succeeded,
     get_command_info,
