+++ e2e_tests/tests/utils.py
@@ -1,29 +1,7 @@
-import json
-import subprocess
-from typing import Any, List
+from typing import Any
 
 from determined.common.api import authentication
 from tests import config as conf
-
-
-def det_cmd(cmd: List[str], **kwargs: Any) -> subprocess.CompletedProcess:
-    return subprocess.run(
-        ["det", "-m", conf.make_master_url()] + cmd,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.PIPE,
-        **kwargs,
-    )
-
-
-def det_cmd_json(cmd: List[str]) -> Any:
-    res = det_cmd(cmd, check=True)
-    return json.loads(res.stdout)
-
-
-def det_cmd_expect_error(cmd: List[str], expected: str) -> None:
-    res = det_cmd(cmd)
-    assert res.returncode != 0
-    assert expected in res.stderr.decode()
 
 
 class CliArgsMock:
